backend/pyproject.toml
[project]
name = "tribe-backend"
version = "0.1.0"
description = "Add your description here"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.118.0",
    "openai>=2.1.0",
    "python-dotenv>=1.1.1",
    "uvicorn>=0.37.0",
]
